moved scanner to User class; added placeShips function to Player class (and its children)
